Move common sync functionality
Create a sync_interface file where common functionality from the different
sync implementations can be imported
