+++ consensus/src/sync/mod.rs
@@ -4,6 +4,7 @@ pub mod light;
 pub mod live;
 pub mod peer_list;
 pub mod pico;
+pub mod sync_interface;
 mod sync_queue;
 pub mod syncer;
 pub mod syncer_proxy;
